test the restructure bold workflow
